+++ internal/provider/schema/job_definition/input_option.go
@@ -23,6 +23,7 @@ func InputOptionSchema() schema.Attribute {
 			"http_input_option":                HttpInputOptionSchema(),
 			"kintone_input_option":             KintoneInputOptionSchema(),
 			"yahoo_ads_api_yss_input_option":   YahooAdsApiYssInputOptionSchema(),
+			"sftp_input_option":                SftpInputOptionSchema(),
 			"hubspot_input_option":             HubspotInputOptionSchema(),
 			"databricks_input_option":          DatabricksInputOptionSchema(),
 		},